requirements.txt
APScheduler==3.7.0
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.2
django-redis==4.12.1
djangorestframework==3.12.4
idna==2.10
pycodestyle==2.7.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
urllib3==1.26.4
psycopg2-binary==2.8.6
gunicorn==20.0.4